backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.25.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
pdfplumber>=0.10.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.1
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.66.0
uvicorn>=0.24.0
gunicorn>=21.2.0
loguru>=0.7.0
pytest>=8.0.0

